+++ src/main/java/team/returm/jobis/domain/student/presentation/StudentController.java
@@ -1,14 +1,9 @@
 package team.returm.jobis.domain.student.presentation;
 
-import team.returm.jobis.domain.student.presentation.dto.request.SendAuthCodeRequest;
 import team.returm.jobis.domain.student.presentation.dto.request.StudentSignUpRequest;
 import team.returm.jobis.domain.student.presentation.dto.request.UpdatePasswordRequest;
-import team.returm.jobis.domain.student.presentation.dto.request.VerifyAuthCodeRequest;
-import team.returm.jobis.domain.student.service.SendPasswordAuthCodeService;
-import team.returm.jobis.domain.student.service.SendSignUpAuthCodeService;
 import team.returm.jobis.domain.student.service.StudentSignUpService;
 import team.returm.jobis.domain.student.service.UpdatePasswordService;
-import team.returm.jobis.domain.student.service.VerifyAuthCodeService;
 import team.returm.jobis.domain.user.presentation.dto.response.TokenResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -26,29 +21,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class StudentController {
 
-    private final SendSignUpAuthCodeService sendSignUpAuthCodeService;
-    private final VerifyAuthCodeService verifyAuthCodeService;
     private final StudentSignUpService studentSignUpService;
-    private final SendPasswordAuthCodeService sendPasswordAuthCodeService;
     private final UpdatePasswordService updatePasswordService;
-
-    @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping("/code/signup")
-    public void sendSignUpCode(@RequestBody @Valid SendAuthCodeRequest request) {
-        sendSignUpAuthCodeService.execute(request);
-    }
-
-    @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping("/code/password")
-    public void sendPasswordCode(@RequestBody @Valid SendAuthCodeRequest request) {
-        sendPasswordAuthCodeService.execute(request);
-    }
-
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PatchMapping("/code")
-    public void verifyCode(@RequestBody @Valid VerifyAuthCodeRequest request) {
-        verifyAuthCodeService.execute(request);
-    }
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("/signup")
